cleaned up some trials
