+++ src/popup/actions/party.ts
@@ -63,13 +63,3 @@ export const joinParty = ({ hash }: any) => {
     });
   };
 };
-
-export const urlChange = ({ newUrl }: any) => {
-  // current issue in this is that if we
-  function changeCurrWin() {
-    chrome.tabs.query({ active: true, currentWindow: true }, tab => {
-      chrome.tabs.update(tab.id, { url: newUrl });
-    });
-  }
-  setTimeout(changeCurrWin, 3000);
-};
